# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Aplikasi PKB",
        "description": "ini adalah dokumentasi API Aplikasi PKB untuk integrasi dengan sistem PKB",
        "termsOfService": "http://lpjk.pu.go.id/pkb/terms/",
        "contact": {
            "email": "admin@lpjk.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://dev.siki.pu.go.id/pkb/api/api/",
            "description": "PKB API host"
        },
        {
            "url": "http://localhost:8000/api/",
            "description": "Local API host"
        }
    ],
    "paths": {
...
        "/v1/kegiatan": {
            "get": {
                "tags": [
                    "Kegiatan"
                ],
                "summary": "mendapatkan data kegiatan",
                "description": "",
                "operationId": "getAuthUser",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "headers": {
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "schema": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error xXx"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Kegiatan"
                ],
                "summary": "membuat data kegiatan baru",
                "operationId": "kegiatan",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nama_kegiatan": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "nama kegiatan"
                                    },
                                    "lokasi_kegiatan": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "www.penyelenggara.com"
                                    },
                                    "unsur": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "U001"
                                    },
                                    "klasifikasi": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "AS"
                                    },
                                    "metode": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "M01"
                                    },
                                    "kode_penyelenggara": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2761"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/kegiatan/{kode}/peserta": {
            "get": {
                "tags": [
                    "Kegiatan"
                ],
                "summary": "mendapatkan data peserta kegiatan",
                "description": "",
                "operationId": "getPeserta",
                "parameters": [
                    {
                        "name": "kode",
                        "in": "path",
                        "description": "kode kegiatan",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "headers": {
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "schema": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error xXx"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Kegiatan"
                ],
                "summary": "membuat data peserta kegiatan baru",
                "operationId": "pesertaKegiatan",
                "parameters": [
                    {
                        "name": "kode",
                        "in": "path",
                        "description": "kode kegiatan",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nik",
                                    "nama_peserta"
                                ],
                                "properties": {
                                    "nik": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "XXXXXXXXXXXXXXXX"
                                    },
                                    "nama_peserta": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "nama peserta"
                                    },
                                    "unsur": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "U001"
                                    },
                                    "metode": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "M01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/unsur": {
            "get": {
                "tags": [
                    "Refrensi"
                ],
                "summary": "Get list of active tests",
                "description": "Returns list of active tests",
                "operationId": "get unsur list",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/v1/klasifikasi": {
            "get": {
                "tags": [
                    "Refrensi"
                ],
                "summary": "Get list of active tests",
                "description": "Returns list of active tests",
                "operationId": "get klasifikasi list",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/v1/metode": {
            "get": {
                "tags": [
                    "Refrensi"
                ],
                "summary": "Get list of active tests",
                "description": "Returns list of active tests",
                "operationId": "get metode list",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "refreshUrl": "http://localhost:8000/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    ]
}